+++ skill/deploy/deploy.properties.sample
@@ -1,0 +1,2 @@
+region=target region name
+function=target lambda function name
